+++ old-ui/app/components/pending-tx.js
@@ -44,7 +44,7 @@ function mapStateToProps (state) {
     unapprovedMsgs: state.metamask.unapprovedMsgs,
     unapprovedPersonalMsgs: state.metamask.unapprovedPersonalMsgs,
     unapprovedTypedMessages: state.metamask.unapprovedTypedMessages,
-    index: state.appState.currentView.context,
+    index: state.appState.currentView.pendingTxIndex || 0,
     warning: state.appState.warning,
     network: state.metamask.network,
     provider: state.metamask.provider,
@@ -519,7 +519,7 @@ PendingTx.prototype.miniAccountPanelForRecipient = function () {
       ])
   } else {
     return h(MiniAccountPanel, {
-      picOrder: 'left',
+      picOrder: 'right',
     }, [
 
       h('span.font-small', {