+++ test/e2e/metamask.spec.js
@@ -7,57 +7,10 @@ const webdriver = require('selenium-webdriver')
 const { By, Key, until } = webdriver
 const { clearField, delay, buildChromeWebDriver, buildFirefoxWebdriver, installWebExt, getExtensionIdChrome, getExtensionIdFirefox } = require('./func')
 
-const accountsMenuSelector = '#app-content > div > div.full-width > div > div:nth-child(2) > span > div'
-const settingsTitleSelector = '#app-content > div > div.app-primary.from-right > div > div.section-title.flex-row.flex-center > h2'
-const deleteImportedAccountTitleSelector = '#app-content > div > div.app-primary.from-left > div > div.section-title.flex-row.flex-center > h2'
-const importedAccountRemoveIconSelector = '#app-content > div > div.full-width > div > div:nth-child(2) > span > div > div > span > div > li:nth-child(4) > div.remove'
-const importedLabelSelector = '#app-content > div > div.full-width > div > div:nth-child(2) > span > div > div > span > div > li:nth-child(4) > div.keyring-label'
-const buttonChangePassword = '#app-content > div > div.app-primary.from-right > div > div.flex-column.flex-justify-center.flex-grow.select-none > div > div:nth-child(10) > button:nth-child(5)'
 let password = '123456789'
+const loader = '#app-content > div > div.full-flex-height > img'
+const { menus, screens, NETWORKS } = require('./elements')
 
-const sandwichMenuSelectors = {
-  Menu: '.sandwich-expando',
-  Settings: '#app-content > div > div:nth-child(3) > span > div > li:nth-child(2)',
-  LogOut: '#app-content > div > div:nth-child(3) > span > div > li:nth-child(3)',
-}
-
-const titlesOfScreensSelectors = {
-  ChangePassword: 'Change Password',
-  Settings: 'Settings',
-}
-const mainScreenSelectors = {
-  buttonBuy: '#app-content > div > div.app-primary.from-right > div > div > div.flex-row > button:nth-child(3)',
-}
-
-const screenChangePassword = {
-  ById: {
-    fieldOldPassword: 'old-password-box',
-    fieldNewPassword: 'new-password-box',
-    fieldConfirmNewPassword: 'password-box-confirm',
-  },
-  ByCss: {
-    buttonNo: '#app-content > div > div.app-primary.from-right > div > div.flex-row.flex-right > button.btn-violet',
-    buttonYes: '#app-content > div > div.app-primary.from-right > div > div.flex-row.flex-right > button:nth-child(2)',
-  },
-  ByClassName: {
-    label: 'confirm-label',
-    arrowLeft: 'fa fa-arrow-left fa-lg cursor-pointer',
-    error: 'error',
-  },
-  labelText: 'Are you sure you want to change the password for unlocking of your wallet?',
-  error: {
-    differ: 'New password should differ from the current one',
-    notLong: 'Password not long enough',
-    dontMatch: 'Passwords don\'t match',
-    incorrectPassword: 'Incorrect password',
-  },
-}
-const screenLock = {
-  fieldPassword: 'password-box',
-  error: 'error',
-  errorText: 'Incorrect password',
-  buttonLogin: 'cursor-pointer',
-}
 
 describe('Metamask popup page', async function () {
   let driver, accountAddress, tokenAddress, extensionId
@@ -100,7 +53,7 @@ describe('Metamask popup page', async function () {
   })
 
   after(async function () {
-    await driver.quit()
+    // await driver.quit()
   })
 
   describe('Setup', async function () {
@@ -115,25 +68,24 @@ describe('Metamask popup page', async function () {
       await delay(300)
       await setProviderType('localhost')
     })
-
   })
 
   describe('Account Creation', async () => {
 
     it('matches Nifty Wallet title', async () => {
       const title = await driver.getTitle()
-      assert.equal(title, 'Nifty Wallet', 'title matches Nifty Wallet')
+      assert.equal(title, screens.TOU.titleText, 'title matches Nifty Wallet')
       await delay(300)
     })
 
     it('show terms of use', async () => {
-      const terms = await driver.findElement(By.css('.terms-header')).getText()
+      const terms = await driver.findElement(By.css(screens.TOU.header)).getText()
       assert.equal(terms, 'Terms of Use', 'shows terms of use')
       delay(300)
     })
 
     it('checks if the TOU button is enabled', async () => {
-      const button = await driver.findElement(By.css('button')).isEnabled()
+      const button = await driver.findElement(By.css(screens.TOU.button)).isEnabled()
       assert.equal(button, true, 'enabled continue button')
       const element = await driver.findElement(By.linkText('Terms of Service'))
       await driver.executeScript('arguments[0].scrollIntoView(true)', element)
@@ -141,14 +93,14 @@ describe('Metamask popup page', async function () {
     })
 
     it('allows the button to be clicked when scrolled to the bottom of TOU', async () => {
-      const button = await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > div.flex-column.flex-center.flex-grow > button'))
+      const button = await driver.findElement(By.css(screens.TOU.button))
       await button.click()
     })
 
     it('accepts password with length of eight', async () => {
-      const passwordBox = await driver.findElement(By.id('password-box'))
-      const passwordBoxConfirm = await driver.findElement(By.id('password-box-confirm'))
-      const button = await driver.findElements(By.css('button'))
+      const passwordBox = await driver.findElement(By.id(screens.create.fieldPassword))
+      const passwordBoxConfirm = await driver.findElement(By.id(screens.create.fieldPasswordConfirm))
+      const button = await driver.findElements(By.css(screens.create.button))
 
       await passwordBox.sendKeys(password)
       await passwordBoxConfirm.sendKeys(password)
@@ -156,54 +108,55 @@ describe('Metamask popup page', async function () {
       await delay(500)
     })
 
-    it('shows value was created and seed phrase', async () => {
+    it('shows vault was created and seed phrase', async () => {
       await delay(300)
-      const element = await driver.findElement(By.css('.twelve-word-phrase'))
+      const element = await driver.findElement(By.css(screens.seedPhrase.fieldPhrase))
       const seedPhrase = await element.getText()
       assert.equal(seedPhrase.split(' ').length, 12)
-      const continueAfterSeedPhrase = await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > button:nth-child(4)'))
-      assert.equal(await continueAfterSeedPhrase.getText(), `I've copied it somewhere safe`)
+      const continueAfterSeedPhrase = await driver.findElement(By.css(screens.seedPhrase.buttonIveCopied))
+      assert.equal(await continueAfterSeedPhrase.getText(), screens.seedPhrase.textButtonIveCopied)
       await continueAfterSeedPhrase.click()
       await delay(300)
     })
 
     it('adds a second account', async function () {
-      await driver.findElement(By.css(accountsMenuSelector)).click()
+      // throw('sf')
+      await driver.findElement(By.css(menus.account.menu)).click()
       await delay(300)
-      await driver.findElement(By.css('#app-content > div > div.full-width > div > div:nth-child(2) > span > div > div > span > div > li:nth-child(3) > span')).click()
+      await driver.findElement(By.css(menus.account.createAccount)).click()
     })
 
     it('shows account address', async function () {
       await delay(300)
-      accountAddress = await driver.findElement(By.css('#app-content > div > div.app-primary.from-left > div > div > div:nth-child(1) > flex-column > div.flex-row > div')).getText()
+      accountAddress = await driver.findElement(By.css(screens.main.address)).getText()
     })
 
     it('logs out of the vault', async () => {
-      await driver.findElement(By.css('.sandwich-expando')).click()
+      await driver.findElement(By.css(menus.sandwich.menu)).click()
       await delay(500)
-      const logoutButton = await driver.findElement(By.css('#app-content > div > div:nth-child(3) > span > div > li:nth-child(3)'))
+      const logoutButton = await driver.findElement(By.css(menus.sandwich.logOut))
       assert.equal(await logoutButton.getText(), 'Log Out')
       await logoutButton.click()
     })
 
     it('accepts account password after lock', async () => {
       await delay(500)
-      await driver.findElement(By.id('password-box')).sendKeys(password)
-      await driver.findElement(By.id('password-box')).sendKeys(Key.ENTER)
+      await driver.findElement(By.id(screens.lock.fieldPassword)).sendKeys(password)
+      await driver.findElement(By.id(screens.lock.fieldPassword)).sendKeys(Key.ENTER)
       await delay(500)
     })
 
     it('shows QR code option', async () => {
       await delay(300)
-      await driver.findElement(By.css('.account-dropdown')).click()
-      await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > div > div:nth-child(1) > flex-column > div.name-label > div > span > div > div > div > li:nth-child(3)')).click()
+      await driver.findElement(By.css(menus.dot.menu)).click()
+      await driver.findElement(By.css(menus.dot.showQRcode)).click()
       await delay(300)
     })
 
     it('checks QR code address is the same as account details address', async () => {
-      const QRaccountAddress = await driver.findElement(By.css('.ellip-address')).getText()
+      const QRaccountAddress = await driver.findElement(By.css(screens.QRcode.address)).getText()
       assert.equal(accountAddress.toLowerCase(), QRaccountAddress)
-      await driver.findElement(By.css('.fa-arrow-left')).click()
+      await driver.findElement(By.css(screens.QRcode.buttonArrow)).click()
       await delay(500)
     })
   })
@@ -219,37 +172,37 @@ describe('Metamask popup page', async function () {
     let fieldOldPassword
     let buttonYes
 
-    describe('check screen "Settings" -> "Change password" ', async () => {
+    describe('Check screen "Settings" -> "Change password" ', async () => {
 
       it('checks if "Change password" button is present and enabled', async () => {
-        await driver.findElement(By.css(sandwichMenuSelectors.Menu)).click()
+        await driver.findElement(By.css(menus.sandwich.menu)).click()
         await delay(500)
-        await driver.findElement(By.css(sandwichMenuSelectors.Settings)).click()
+        await driver.findElement(By.css(menus.sandwich.settings)).click()
         await delay(500)
-        const buttons = await driver.findElements(By.css(buttonChangePassword))
+        const buttons = await driver.findElements(By.css(screens.settings.buttons.changePassword))
         assert.equal(buttons.length, 1, 'Button "Change password" is not present')
         assert.equal(await buttons[0].isEnabled(), true, 'Button "Change password" is disabled')
       })
 
       it('screen contains correct title', async () => {
-        const button = await driver.findElement(By.css(buttonChangePassword))
+        const button = await driver.findElement(By.css(screens.settings.buttons.changePassword))
         await button.click()
         const title = await driver.findElement(By.className('page-subtitle'))
-        assert.equal(await title.getText(), titlesOfScreensSelectors.ChangePassword, '"Change password" screen contains incorrect title')
+        assert.equal(await title.getText(), screens.changePassword.titleText, '"Change password" screen contains incorrect title')
       })
 
       it('screen contains correct label', async () => {
-        const labels = await driver.findElements(By.className(screenChangePassword.ByClassName.label))
+        const labels = await driver.findElements(By.className(screens.changePassword.ByClassName.label))
         assert.equal(labels.length, 1, 'screen "Change password" doesn\'t contain label')
-        assert.equal(await labels[0].getText(), screenChangePassword.labelText, 'label contains incorrect title')
+        assert.equal(await labels[0].getText(), screens.changePassword.labelText, 'label contains incorrect title')
       })
 
       it('clicking the button "No" bring back to "Setting" screen ', async () => {
-        const button = await driver.findElement(By.css(screenChangePassword.ByCss.buttonNo))
+        const button = await driver.findElement(By.css(screens.changePassword.ByCss.buttonNo))
         await button.click()
-        const title = await driver.findElement(By.css(settingsTitleSelector))
-        assert.equal(await title.getText(), titlesOfScreensSelectors.Settings, 'button "No" doesnt open settings screen')
-        const buttonChangePass = await driver.findElement(By.css(buttonChangePassword))
+        const title = await driver.findElement(By.css(screens.settings.title))
+        assert.equal(await title.getText(), screens.settings.titleText, 'button "No" doesnt open settings screen')
+        const buttonChangePass = await driver.findElement(By.css(screens.settings.buttons.changePassword))
         await buttonChangePass.click()
       })
     })
@@ -257,20 +210,20 @@ describe('Metamask popup page', async function () {
     describe('Validation of errors ', async () => {
 
       before(async () => {
-        fieldOldPassword = await driver.findElement(By.id(screenChangePassword.ById.fieldOldPassword))
+        fieldOldPassword = await driver.findElement(By.id(screens.changePassword.ById.fieldOldPassword))
         await fieldOldPassword.sendKeys(password)
-        fieldNewPassword = await driver.findElement(By.id(screenChangePassword.ById.fieldNewPassword))
-        fieldConfirmNewPassword = await driver.findElement(By.id(screenChangePassword.ById.fieldConfirmNewPassword))
-        buttonYes = await driver.findElement(By.css(screenChangePassword.ByCss.buttonYes))
+        fieldNewPassword = await driver.findElement(By.id(screens.changePassword.ById.fieldNewPassword))
+        fieldConfirmNewPassword = await driver.findElement(By.id(screens.changePassword.ById.fieldConfirmNewPassword))
+        buttonYes = await driver.findElement(By.css(screens.changePassword.ByCss.buttonYes))
       })
 
       it('error if new password shorter than 8 digits', async () => {
         await fieldNewPassword.sendKeys(newPassword.short)
         await fieldConfirmNewPassword.sendKeys(newPassword.short)
         await buttonYes.click()
-        const errors = await driver.findElements(By.className(screenChangePassword.ByClassName.error))
+        const errors = await driver.findElements(By.className(screens.changePassword.ByClassName.error))
         assert.equal(errors.length > 0, true, 'error isn\'t displayed')
-        assert.equal(await errors[0].getText(), screenChangePassword.error.notLong, 'Error\'s text incorrect')
+        assert.equal(await errors[0].getText(), screens.changePassword.error.notLong, 'Error\'s text incorrect')
       })
 
       it('error if new password  doesn\'t match confirmation', async () => {
@@ -279,9 +232,9 @@ describe('Metamask popup page', async function () {
         await fieldNewPassword.sendKeys(newPassword.correct)
         await fieldConfirmNewPassword.sendKeys(newPassword.incorrect)
         await buttonYes.click()
-        const errors = await driver.findElements(By.className(screenChangePassword.ByClassName.error))
+        const errors = await driver.findElements(By.className(screens.changePassword.ByClassName.error))
         assert.equal(errors.length > 0, true, 'error isn\'t displayed')
-        assert.equal(await errors[0].getText(), screenChangePassword.error.dontMatch, 'Error\'s text incorrect')
+        assert.equal(await errors[0].getText(), screens.changePassword.error.dontMatch, 'Error\'s text incorrect')
       })
 
       it('error if new password match old password', async () => {
@@ -290,18 +243,18 @@ describe('Metamask popup page', async function () {
         await fieldNewPassword.sendKeys(password)
         await fieldConfirmNewPassword.sendKeys(password)
         await buttonYes.click()
-        const errors = await driver.findElements(By.className(screenChangePassword.ByClassName.error))
+        const errors = await driver.findElements(By.className(screens.changePassword.ByClassName.error))
         assert.equal(errors.length > 0, true, 'error isn\'t displayed')
-        assert.equal(await errors[0].getText(), screenChangePassword.error.differ, 'Error\'s text incorrect')
+        assert.equal(await errors[0].getText(), screens.changePassword.error.differ, 'Error\'s text incorrect')
       })
 
       it.skip('error if old password incorrect, https://github.com/poanetwork/metamask-extension/issues/86 ', async () => {
         await clearField(fieldOldPassword)
         await fieldOldPassword.sendKeys(newPassword.incorrect)
         await buttonYes.click()
-        const errors = await driver.findElements(By.className(screenChangePassword.ByClassName.error))
+        const errors = await driver.findElements(By.className(screens.changePassword.ByClassName.error))
         assert.equal(errors.length > 0, true, 'error isn\'t displayed')
-        assert.equal(await errors[0].getText(), screenChangePassword.error.incorrectPassword, 'Error\'s text incorrect')
+        assert.equal(await errors[0].getText(), screens.changePassword.error.incorrectPassword, 'Error\'s text incorrect')
       })
 
       it('no errors if old, new, confirm new passwords are correct; user can change password', async () => {
@@ -314,8 +267,8 @@ describe('Metamask popup page', async function () {
         await fieldConfirmNewPassword.sendKeys(newPassword.correct)
         await buttonYes.click()
 
-        await driver.wait(until.elementLocated(By.css(buttonChangePassword)))
-        const buttons = await driver.findElements(By.css(buttonChangePassword))
+        await driver.wait(until.elementLocated(By.css(screens.settings.buttons.changePassword)))
+        const buttons = await driver.findElements(By.css(screens.settings.buttons.changePassword))
         assert.equal(buttons.length, 1, 'Button "Change password" is not present')
         assert.equal(await buttons[0].isEnabled(), true, 'Button "Change password" is disabled')
       })
@@ -324,31 +277,31 @@ describe('Metamask popup page', async function () {
     describe('Check if new password is accepted', async () => {
 
       it('user can log out', async () => {
-        await driver.findElement(By.css(sandwichMenuSelectors.Menu)).click()
+        await driver.findElement(By.css(menus.sandwich.menu)).click()
         await delay(500)
-        await driver.wait(until.elementLocated(By.css(sandwichMenuSelectors.LogOut)))
-        const itemLogOut = await driver.findElement(By.css(sandwichMenuSelectors.LogOut))
+        await driver.wait(until.elementLocated(By.css(menus.sandwich.logOut)))
+        const itemLogOut = await driver.findElement(By.css(menus.sandwich.logOut))
         await driver.wait(until.elementIsVisible(itemLogOut))
         itemLogOut.click()
-        await driver.wait(until.elementLocated(By.id(screenLock.fieldPassword)))
-        const fields = await driver.findElements(By.id(screenLock.fieldPassword))
+        await driver.wait(until.elementLocated(By.id(screens.lock.fieldPassword)))
+        const fields = await driver.findElements(By.id(screens.lock.fieldPassword))
         assert.equal(fields.length, 1, 'password box isn\'t present after logout')
       })
       it.skip('can\'t login with old password', async () => {
-        const field = await driver.findElement(By.id(screenLock.fieldPassword))
+        const field = await driver.findElement(By.id(screens.lock.fieldPassword))
         await field.sendKeys(password)
-        await driver.findElement(By.className(screenLock.buttonLogin)).click()
-        const errors = await driver.findElements(By.className(screenLock.error))
+        await driver.findElement(By.className(screens.lock.buttonLogin)).click()
+        const errors = await driver.findElements(By.className(screens.lock.error))
         assert.equal(errors.length, 1, 'error isn\'t displayed if password incorrect')
-        assert.equal(await errors[0].getText(), screenLock.errorText, 'error\'s text incorrect')
+        assert.equal(await errors[0].getText(), screens.lock.errorText, 'error\'s text incorrect')
       })
       it('accepts new password after lock', async () => {
-        const field = await driver.findElement(By.id(screenLock.fieldPassword))
+        const field = await driver.findElement(By.id(screens.lock.fieldPassword))
         await field.sendKeys(newPassword.correct)
-        await driver.findElement(By.className(screenLock.buttonLogin)).click()
+        await driver.findElement(By.className(screens.lock.buttonLogin)).click()
 
-        await driver.wait(until.elementLocated(By.css(mainScreenSelectors.buttonBuy)))
-        const buttons = await driver.findElements(By.css(mainScreenSelectors.buttonBuy))
+        await driver.wait(until.elementLocated(By.css(screens.main.buttons.buy)))
+        const buttons = await driver.findElements(By.css(screens.main.buttons.buy))
         assert.equal(buttons.length, 1, 'main screen isn\'t displayed')
         password = newPassword.correct
       })
@@ -358,67 +311,67 @@ describe('Metamask popup page', async function () {
   describe('Import Account', () => {
 
     it('opens import account menu', async function () {
-      await driver.wait(until.elementLocated(By.css(accountsMenuSelector)))
-      await driver.findElement(By.css(accountsMenuSelector)).click()
+      await driver.wait(until.elementLocated(By.css(menus.account.menu)))
+      await driver.findElement(By.css(menus.account.menu)).click()
       await delay(500)
-      await driver.findElement(By.css('#app-content > div > div.full-width > div > div:nth-child(2) > span > div > div > span > div > li:nth-child(5) > span')).click()
+      await driver.findElement(By.css(menus.account.import)).click()
       await delay(500)
-      const importAccountTitle = await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > div:nth-child(2) > div.flex-row.flex-center > h2'))
-      assert.equal(await importAccountTitle.getText(), 'Import Accounts')
+      const importAccountTitle = await driver.findElement(By.css(screens.importAccounts.title))
+      assert.equal(await importAccountTitle.getText(), screens.importAccounts.textTitle)
     })
 
     it('imports account', async function () {
-      const privateKeyBox = await driver.findElement(By.css('#private-key-box'))
-      const importButton = await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > div:nth-child(2) > div:nth-child(4) > button'))
+      const privateKeyBox = await driver.findElement(By.css(screens.importAccounts.fieldPrivateKey))
+      const importButton = await driver.findElement(By.css(screens.importAccounts.buttonImport))
       privateKeyBox.sendKeys('c6b81c1252415d1acfda94474ab8f662a44c045f96749c805ff12a6074081586')// demo private key
       importButton.click()
       await delay(200)
       // check, that account is added
-      await driver.findElement(By.css(accountsMenuSelector)).click()
+      await driver.findElement(By.css(menus.account.menu)).click()
       await delay(500)
-      const importedLabel = await driver.findElement(By.css(importedLabelSelector))
+      const importedLabel = await driver.findElement(By.css(menus.account.labelImported))
       assert.equal(await importedLabel.getText(), 'IMPORTED')
     })
 
     it('opens delete imported account screen', async function () {
-      await driver.findElement(By.css(importedAccountRemoveIconSelector)).click()
+      await driver.findElement(By.css(menus.account.delete)).click()
       await delay(200)
-      const deleteImportedAccountTitle = await driver.findElement(By.css(deleteImportedAccountTitleSelector))
-      assert.equal(await deleteImportedAccountTitle.getText(), 'Delete Imported Account')
+      const deleteImportedAccountTitle = await driver.findElement(By.css(screens.deleteImportedAccount.title))
+      assert.equal(await deleteImportedAccountTitle.getText(), screens.deleteImportedAccount.titleText)
     })
 
     it('doesn\'t remove imported account with \'No\' button', async function () {
-      const NoButton = await driver.findElement(By.css('#app-content > div > div.app-primary.from-left > div > div.flex-row.flex-right > button.btn-violet'))
+      const NoButton = await driver.findElement(By.css(screens.deleteImportedAccount.buttons.no))
       NoButton.click()
       await delay(500)
-      const settingsTitle = await driver.findElement(By.css(settingsTitleSelector))
+      const settingsTitle = await driver.findElement(By.css(screens.settings.title))
       assert.equal(await settingsTitle.getText(), 'Settings')
 
       // check, that imported account still exists
-      await driver.findElement(By.css(accountsMenuSelector)).click()
+      await driver.findElement(By.css(menus.account.menu)).click()
       await delay(500)
-      const importedLabel = await driver.findElement(By.css(importedLabelSelector))
+      const importedLabel = await driver.findElement(By.css(menus.account.labelImported))
       assert.equal(await importedLabel.getText(), 'IMPORTED')
     })
 
     it('opens delete imported account screen again', async function () {
-      await driver.findElement(By.css(importedAccountRemoveIconSelector)).click()
+      await driver.findElement(By.css(menus.account.delete)).click()
       await delay(500)
     })
 
     it('removes imported account with \'Yes\' button', async function () {
-      const YesButton = await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > div.flex-row.flex-right > button:nth-child(2)'))
+      const YesButton = await driver.findElement(By.css(screens.deleteImportedAccount.buttons.yes))
       YesButton.click()
       await delay(500)
-      const settingsTitle = await driver.findElement(By.css(settingsTitleSelector))
+      const settingsTitle = await driver.findElement(By.css(screens.settings.title))
       assert.equal(await settingsTitle.getText(), 'Settings')
 
       // check, that imported account is removed
-      await driver.findElement(By.css(accountsMenuSelector)).click()
+      await driver.findElement(By.css(menus.account.menu)).click()
       await delay(500)
-      const importedAccounts = await driver.findElements(By.css(importedLabelSelector))
+      const importedAccounts = await driver.findElements(By.css(menus.account.labelImported))
       assert.ok(importedAccounts.length === 0)
-      await driver.findElement(By.css(accountsMenuSelector)).click()
+      await driver.findElement(By.css(menus.account.menu)).click()
       await delay(500)
     })
   })
@@ -426,60 +379,60 @@ describe('Metamask popup page', async function () {
   describe('Import Ganache seed phrase', function () {
 
     it('logs out', async function () {
-      await driver.findElement(By.css('.sandwich-expando')).click()
+      await driver.findElement(By.css(menus.sandwich.menu)).click()
       await delay(200)
-      const logOut = await driver.findElement(By.css('#app-content > div > div:nth-child(3) > span > div > li:nth-child(3)'))
-      assert.equal(await logOut.getText(), 'Log Out')
+      const logOut = await driver.findElement(By.css(menus.sandwich.logOut))
+      assert.equal(await logOut.getText(), menus.sandwich.textLogOut)
       await logOut.click()
       await delay(300)
     })
 
     it('restores from seed phrase', async function () {
-      const restoreSeedLink = await driver.findElement(By.css('#app-content > div > div.app-primary.from-left > div > div.flex-row.flex-center.flex-grow > p'))
-      assert.equal(await restoreSeedLink.getText(), 'Restore from seed phrase')
+      const restoreSeedLink = await driver.findElement(By.css(screens.lock.linkRestore))
+      assert.equal(await restoreSeedLink.getText(), screens.lock.linkRestoreText)
       await restoreSeedLink.click()
       await delay(100)
     })
 
     it('adds seed phrase', async function () {
       const testSeedPhrase = 'phrase upgrade clock rough situate wedding elder clever doctor stamp excess tent'
-      const seedTextArea = await driver.findElement(By.css('#app-content > div > div.app-primary.from-left > div > div.initialize-screen.flex-column.flex-center.flex-grow > textarea'))
+      const seedTextArea = await driver.findElement(By.css(screens.restoreVault.textArea))
       await seedTextArea.sendKeys(testSeedPhrase)
 
-      await driver.findElement(By.id('password-box')).sendKeys('123456789')
-      await driver.findElement(By.id('password-box-confirm')).sendKeys('123456789')
-      await driver.findElement(By.css('#app-content > div > div.app-primary.from-left > div > div.initialize-screen.flex-column.flex-center.flex-grow > div > button:nth-child(2)')).click()
+      await driver.findElement(By.id(screens.restoreVault.fieldPassword)).sendKeys(password)
+      await driver.findElement(By.id(screens.restoreVault.fieldPasswordConfirm)).sendKeys(password)
+      await driver.findElement(By.css(screens.restoreVault.buttos.ok)).click()
       await delay(500)
     })
 
     it('balance renders', async function () {
       await delay(200)
-      const balance = await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > div > div.flex-row > div.ether-balance.ether-balance-amount > div > div > div:nth-child(1) > div:nth-child(1)'))
+      const balance = await driver.findElement(By.css(screens.main.balance))
       assert.equal(await balance.getText(), '100.000')
       await delay(200)
     })
 
     it('sends transaction', async function () {
-      const sendButton = await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > div > div.flex-row > button:nth-child(4)'))
-      assert.equal(await sendButton.getText(), 'Send')
+      const sendButton = await driver.findElement(By.css(screens.main.buttons.send))
+      assert.equal(await sendButton.getText(), screens.main.buttons.sendText)
       await sendButton.click()
       await delay(200)
     })
 
     it('adds recipient address and amount', async function () {
-      const sendTranscationScreen = await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > h3:nth-child(2)')).getText()
-      assert.equal(sendTranscationScreen, 'Send Transaction')
-      const inputAddress = await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > section:nth-child(3) > div > input'))
-      const inputAmmount = await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > section:nth-child(4) > input'))
+      const sendTranscationScreen = await driver.findElement(By.css(screens.sendTransaction.title)).getText()
+      assert.equal(sendTranscationScreen, screens.sendTransaction.titleText)
+      const inputAddress = await driver.findElement(By.css(screens.sendTransaction.fields.address))
+      const inputAmmount = await driver.findElement(By.css(screens.sendTransaction.fields.amount))
       await inputAddress.sendKeys('0x2f318C334780961FB129D2a6c30D0763d9a5C970')
       await inputAmmount.sendKeys('10')
-      await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > section:nth-child(4) > button')).click()
+      await driver.findElement(By.css(screens.sendTransaction.buttonNext)).click()
       await delay(300)
     })
 
     it('confirms transaction', async function () {
       await delay(300)
-      const bySubmitButton = By.css('#pending-tx-form > div.flex-row.flex-space-around.conf-buttons > input')
+      const bySubmitButton = By.css(screens.confirmTransaction.buttons.submit)
       const submitButton = await driver.wait(until.elementLocated(bySubmitButton))
 
       submitButton.click()
@@ -488,7 +441,7 @@ describe('Metamask popup page', async function () {
     })
 
     it('finds the transaction in the transactions list', async function () {
-      const tranasactionAmount = await driver.findElement(By.css('#app-content > div > div.app-primary.from-left > div > section > section > div > div > div > div.ether-balance.ether-balance-amount > div > div > div > div:nth-child(1)'))
+      const tranasactionAmount = await driver.findElement(By.css(screens.main.transactionList))
       assert.equal(await tranasactionAmount.getText(), '10.0')
     })
   })
@@ -563,7 +516,7 @@ describe('Metamask popup page', async function () {
     })
 
     it('checks add token screen rendered', async function () {
-      const addTokenScreen = await driver.findElement(By.css(settingsTitleSelector))
+      const addTokenScreen = await driver.findElement(By.css(screens.settings.title))
       assert.equal(await addTokenScreen.getText(), 'Add Token')
     })
 
@@ -576,23 +529,119 @@ describe('Metamask popup page', async function () {
     })
 
     it('checks the token balance', async function () {
-      const tokenBalance = await driver.findElement(By.css('#app-content > div > div.app-primary.from-left > div > section > div.full-flex-height > ol > li:nth-child(2) > h3'))
+      const tokenBalance = await driver.findElement(By.css(screens.main.tokens.balance))
       assert.equal(await tokenBalance.getText(), '100 TST')
+    })
+  })
+
+  describe('Check support of token per network basis ', async function () {
+
+    describe('Token should be displayed only for network, where it was added ', async function () {
+
+      it('token should not  be displayed in POA network', async function () {
+        await setProvider(NETWORKS.POA)
+        assert.equal(await assertTokensNotDisplayed(), true, 'tokens are displayed')
+      })
+
+      it('token should not  be displayed in SOKOL network', async function () {
+        await setProvider(NETWORKS.SOKOL)
+        assert.equal(await assertTokensNotDisplayed(), true, 'tokens are displayed')
+      })
+
+      it('token should not  be displayed in MAINNET network', async function () {
+        await setProvider(NETWORKS.MAINNET)
+        assert.equal(await assertTokensNotDisplayed(), true, 'tokens are displayed')
+      })
+
+      it('token should not  be displayed in ROPSTEN network', async function () {
+        await setProvider(NETWORKS.ROPSTEN)
+        assert.equal(await assertTokensNotDisplayed(), true, 'tokens are displayed')
+      })
+
+      it('token should not  be displayed in KOVAN network', async function () {
+        await setProvider(NETWORKS.KOVAN)
+        assert.equal(await assertTokensNotDisplayed(), true, 'tokens are displayed')
+      })
+
+      it('token should not  be displayed in RINKEBY network', async function () {
+        await setProvider(NETWORKS.RINKEBY)
+        assert.equal(await assertTokensNotDisplayed(), true, 'tokens are displayed')
+      })
+    })
+
+    describe('Add token with the same address to each network  ', async function () {
+
+      const tokenName = 'DVT'
+      const tokenDecimals = '13'
+
+      it('adds token with  the same address to POA network', async function () {
+        await setProvider(NETWORKS.POA)
+        await addToken(tokenAddress, tokenName, tokenDecimals)
+        const tokenBalance = await await waitUntilShowUp(By.css(screens.main.tokens.balance))
+        assert.notEqual(await tokenBalance.getText(), '')
+      })
+
+      it('adds token with  the same address to SOKOL network', async function () {
+        await setProvider(NETWORKS.SOKOL)
+        await addToken(tokenAddress, tokenName, tokenDecimals)
+        const tokenBalance = await await waitUntilShowUp(By.css(screens.main.tokens.balance))
+        assert.notEqual(await tokenBalance.getText(), '')
+      })
+
+      it('adds token with  the same address to MAINNET network', async function () {
+        await setProvider(NETWORKS.MAINNET)
+        await addToken(tokenAddress, tokenName, tokenDecimals)
+        const tokenBalance = await await waitUntilShowUp(By.css(screens.main.tokens.balance))
+        assert.notEqual(await tokenBalance.getText(), '')
+      })
+
+      it('adds token with  the same address to ROPSTEN network', async function () {
+        await setProvider(NETWORKS.ROPSTEN)
+        await addToken(tokenAddress, tokenName, tokenDecimals)
+        const tokenBalance = await await waitUntilShowUp(By.css(screens.main.tokens.balance))
+        assert.notEqual(await tokenBalance.getText(), '')
+      })
+
+      it('adds token with  the same address to KOVAN network', async function () {
+        await setProvider(NETWORKS.KOVAN)
+        await addToken(tokenAddress, tokenName, tokenDecimals)
+        const tokenBalance = await await waitUntilShowUp(By.css(screens.main.tokens.balance))
+        assert.notEqual(await tokenBalance.getText(), '')
+      })
+
+      it('adds token with  the same address to RINKEBY network', async function () {
+        await setProvider(NETWORKS.RINKEBY)
+        await addToken(tokenAddress, tokenName, tokenDecimals)
+        const tokenBalance = await await waitUntilShowUp(By.css(screens.main.tokens.balance))
+        assert.notEqual(await tokenBalance.getText(), '')
+      })
+
+      it('token still should be displayed in LOCALHOST network', async function () {
+        await setProvider(NETWORKS.LOCALHOST)
+        await waitUntilDisappear(By.css(screens.main.tokens.amount))
+        assert.notEqual(await waitUntilShowUp(By.css(screens.main.tokens.amount)), false, 'App is frozen')
+        const tokens = await driver.findElements(By.css(screens.main.tokens.amount))
+        assert.equal(tokens.length, 1, '\'Tokens\' section doesn\'t contain field with amount of tokens')
+        assert.equal(await tokens[0].getText(), screens.main.tokens.textYouOwn1token, 'Token isn\'t displayed')
+      })
     })
   })
 
   describe('Remove Token', function () {
     it('navigates to the remove token screen and goes back', async function () {
       // Click to remove first token
+      await setProvider(NETWORKS.LOCALHOST)
       const removeTokenButton = await driver.findElement(By.css('#app-content > div > div.app-primary.from-left > div > section > div.full-flex-height > ol > li:nth-child(2) > .trash'))
       await removeTokenButton.click()
-      const removeTokenTitle = await driver.findElement(By.css(settingsTitleSelector))
+      const removeTokenTitle = await driver.findElement(By.css(screens.settings.title))
 
       // Check that the correct page is opened
       assert.equal(await removeTokenTitle.getText(), 'Remove Token')
 
       // Go back
-      await driver.findElement(By.css('.fa-arrow-left')).click()
+      await delay(500)
+      await driver.findElement(By.className('fa fa-arrow-left fa-lg cursor-pointer')).click()
+
       await delay(300)
 
       // Check that the token was not deleted
@@ -600,11 +649,11 @@ describe('Metamask popup page', async function () {
       assert.equal(tokens.length, 1, 'There should be 1 token')
     })
 
-    it('navigates to the remove token screen and removes the token', async function () {
+    it('navigates to the remove token screen and removes the token from LOCALHOST', async function () {
       // Click to remove first token
       const removeTokenButton = await driver.findElement(By.css('#app-content > div > div.app-primary.from-left > div > section > div.full-flex-height > ol > li:nth-child(2) > .trash'))
       await removeTokenButton.click()
-      const removeTokenTitle = await driver.findElement(By.css(settingsTitleSelector))
+      const removeTokenTitle = await driver.findElement(By.css(screens.settings.title))
 
       // Check that the correct page is opened
       assert.equal(await removeTokenTitle.getText(), 'Remove Token')
@@ -616,50 +665,176 @@ describe('Metamask popup page', async function () {
       await delay(300)
 
       // Check that the token was deleted
-      const tokens = await driver.findElements(By.css('#app-content > div > div.app-primary.from-left > div > section > div.full-flex-height > ol > li'))
-      assert.equal(tokens.length, 0, 'There should be no tokens')
+      assert.equal(await assertTokensNotDisplayed(), true, 'tokens are displayed')
+    })
+
+    it('check if token was removed from SOKOL network', async function () {
+      await setProvider(NETWORKS.SOKOL)
+      assert.equal(await assertTokensNotDisplayed(), true, 'tokens are displayed')
+    })
+
+    it('check if token was removed from KOVAN network', async function () {
+      await setProvider(NETWORKS.KOVAN)
+      assert.equal(await assertTokensNotDisplayed(), true, 'tokens are displayed')
+    })
+
+    it('check if token was removed from ROPSTEN network', async function () {
+      await setProvider(NETWORKS.ROPSTEN)
+      assert.equal(await assertTokensNotDisplayed(), true, 'tokens are displayed')
+    })
+
+    it('check if token was removed from MAINNET network', async function () {
+      await setProvider(NETWORKS.MAINNET)
+      assert.equal(await assertTokensNotDisplayed(), true, 'tokens are displayed')
+    })
+
+    it('check if token was removed from POA network', async function () {
+      await setProvider(NETWORKS.POA)
+      assert.equal(await assertTokensNotDisplayed(), true, 'tokens are displayed')
+    })
+
+    it('check if token was removed from RINKEBY network', async function () {
+      await setProvider(NETWORKS.RINKEBY)
+      assert.equal(await assertTokensNotDisplayed(), true, 'tokens are displayed')
     })
   })
 
   describe('Custom Rpc', function () {
     it('switches to settings screen', async function () {
-      await driver.findElement(By.css('.sandwich-expando')).click()
+      await driver.findElement(By.css(menus.sandwich.menu)).click()
       await delay(200)
-      const settings = await driver.findElement(By.css('#app-content > div > div:nth-child(3) > span > div > li:nth-child(2)'))
-      assert.equal(await settings.getText(), 'Settings')
+      const settings = await driver.findElement(By.css(menus.sandwich.settings))
+      assert.equal(await settings.getText(), menus.sandwich.textSettings)
       await settings.click()
       await delay(300)
     })
 
     it('add custom rpc', async function () {
       const customUrl = 'http://test.com'
-      const input = await driver.findElement(By.css('#new_rpc'))
+      const input = await driver.findElement(By.css(screens.settings.fieldNewRPC))
       input.sendKeys(customUrl)
-      await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > div.flex-column.flex-justify-center.flex-grow.select-none > div > div:nth-child(2) > button')).click()
+      await driver.findElement(By.css(screens.settings.buttonSave)).click()
       if (process.env.SELENIUM_BROWSER === 'firefox') {
         input.sendKeys(Key.ENTER)
       }
       await delay(400)
-      const customUrlElement = await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > div.flex-column.flex-justify-center.flex-grow.select-none > div > div:nth-child(1) > span:nth-child(2)'))
+      const customUrlElement = await driver.findElement(By.css(screens.settings.customUrl))
       assert.equal(await customUrlElement.getText(), customUrl)
     })
 
     it('delete custom rpc', async function () {
-      await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > div.flex-column.flex-justify-center.flex-grow.select-none > div > div:nth-child(1) > button')).click()
+      await driver.findElement(By.css(screens.settings.buttons.delete)).click()
       await delay(300)
-      const titleConfirmPage = await driver.findElement(By.css(settingsTitleSelector))
+      const titleConfirmPage = await driver.findElement(By.css(screens.settings.title))
       assert.equal(await titleConfirmPage.getText(), 'Delete Custom RPC')
-      const yesButton = await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > div.flex-row.flex-right > button:nth-child(2)'))
+      const yesButton = await driver.findElement(By.css(screens.deleteCustomRPC.buttons.yes))
       assert.equal(await yesButton.getText(), 'Yes')
       await yesButton.click()
       await delay(300)
-      const urlElement = await driver.findElement(By.css('#app-content > div > div.app-primary.from-right > div > div.flex-column.flex-justify-center.flex-grow.select-none > div > div:nth-child(1) > span:nth-child(2)'))
+      const urlElement = await driver.findElement(By.css(screens.settings.currentNetwork))
       assert.equal(await urlElement.getText(), 'POA Network')
     })
   })
 
   async function setProviderType (type) {
     await driver.executeScript('window.metamask.setProviderType(arguments[0])', type)
+  }
+
+  async function setProvider (network) {
+    await driver.findElement(By.className(screens.main.network)).click()
+    let counter
+    switch (network) {
+      case NETWORKS.POA:
+        counter = 0
+        break
+      case NETWORKS.SOKOL:
+        counter = 1
+        break
+      case NETWORKS.MAINNET:
+        counter = 2
+        break
+      case NETWORKS.ROPSTEN:
+        counter = 3
+        break
+      case NETWORKS.KOVAN:
+        counter = 4
+        break
+      case NETWORKS.RINKEBY:
+        counter = 5
+        break
+      case NETWORKS.LOCALHOST:
+        counter = 6
+        break
+      case NETWORKS.CUSTOM:
+        counter = 7
+        break
+      default:
+        counter = 6
+    }
+    await driver.executeScript("document.getElementsByClassName('dropdown-menu-item')[" + counter + '].click();')
+  }
+
+  async function waitUntilDisappear (by, Twait) {
+    if (Twait === undefined) Twait = 10
+    do {
+      await delay(100)
+      if (!await isElementDisplayed(by)) return true
+
+    } while (Twait-- > 0)
+    return false
+  }
+
+  async function waitUntilShowUp (by, Twait) {
+    if (Twait === undefined) Twait = 2000
+    do {
+      await delay(100)
+      if (await isElementDisplayed(by)) return await driver.findElement(by)
+    } while (Twait-- > 0)
+    return false
+  }
+
+  async function isElementDisplayed (by) {
+    try {
+      return await driver.findElement(by).isDisplayed()
+    } catch (err) {
+      return false
+    }
+  }
+
+  async function assertTokensNotDisplayed () {
+    try {
+      await delay(800)
+      await waitUntilDisappear(By.css(loader))
+      assert.notEqual(await waitUntilShowUp(By.css(screens.main.tokens.amount)), false, 'App is frozen')
+      // Check tokens title
+      const tokensStatus = await driver.findElements(By.css(screens.main.tokens.amount))
+      assert.equal(tokensStatus.length, 1, '\'Tokens\' section doesn\'t contain field with amount of tokens')
+      assert.equal(await tokensStatus[0].getText(), screens.main.tokens.textNoTokens, 'Unexpected token presents')
+      // Check if token presents
+      const tokens = await driver.findElements(By.css(screens.main.tokens.token))
+      assert.equal(tokens.length, 0, 'Unexpected token presents')
+      return true
+    } catch (err) {
+      return err
+    }
+  }
+
+  async function addToken (tokenAddress, tokenName, tokenDecimals) {
+    try {
+      const button = await waitUntilShowUp(By.css(screens.main.tokens.buttonAdd))
+      await button.click()
+      const field = await waitUntilShowUp(By.css(screens.addToken.fieldTokenContractAddress))
+      await field.sendKeys(tokenAddress)
+      await delay(500)
+      await driver.findElement(By.css(screens.addToken.fieldTokenSymbol)).sendKeys(tokenName)
+      await delay(500)
+      await driver.findElement(By.css(screens.addToken.fieldDecimals)).sendKeys(tokenDecimals)
+      const buttonAdd = await waitUntilShowUp(By.css(screens.addToken.buttonAdd))
+      await buttonAdd.click()
+      return true
+    } catch (err) {
+      return false
+    }
   }
 
   async function checkBrowserForConsoleErrors () {